+++ ricaun.RevitTest.TestAdapter/TestExecutor.cs
@@ -66,6 +66,8 @@ namespace ricaun.RevitTest.TestAdapter
         {
         }
 
+        private static bool IsLogDebug => System.Diagnostics.Debugger.IsAttached || AdapterSettings.Settings.NUnit.Verbosity > 2;
+
         /// <summary>
         /// Run Tests using RevitTest.Console -t
         /// </summary>
@@ -111,8 +113,8 @@ namespace ricaun.RevitTest.TestAdapter
 
                     if (item.StartsWith("{\"FileName"))
                     {
-                        if (System.Diagnostics.Debugger.IsAttached)
-                            AdapterLogger.Logger.Debug($"OutputConsole: Debug: {item.Trim()}");
+                        if (IsLogDebug)
+                            AdapterLogger.Logger.Debug($"OutputConsole: DEBUG: {item.Trim()}");
 
                         var testAssembly = item.Deserialize<TestAssemblyModel>();
 
@@ -126,8 +128,8 @@ namespace ricaun.RevitTest.TestAdapter
                     }
                     else if (item.StartsWith("{\"Name"))
                     {
-                        if (System.Diagnostics.Debugger.IsAttached)
-                            AdapterLogger.Logger.Debug($"OutputConsole: Debug: {item.Trim()}");
+                        if (IsLogDebug)
+                            AdapterLogger.Logger.Debug($"OutputConsole: DEBUG: {item.Trim()}");
 
                         if (item.Deserialize<TestModel>() is TestModel testModel)
                         {
@@ -141,8 +143,8 @@ namespace ricaun.RevitTest.TestAdapter
                     }
                     else
                     {
-                        if (System.Diagnostics.Debugger.IsAttached)
-                            AdapterLogger.Logger.Debug($"OutputConsole: Debug: {item}");
+                        if (IsLogDebug)
+                            AdapterLogger.Logger.Debug($"OutputConsole: DEBUG: {item}");
                     }
                 };